feat: Add install, show, and params commands
This commit adds the following commands to the CLI:
- install: Installs something
- show: Displays data and temperature/humidity data from YaoRozu Farm
- params: Handles parameters

The commands are implemented in separate files under the 'commands' directory.
