+++ commands/params.js
@@ -1,0 +1,8 @@
+import { Command } from 'commander';
+
+export const paramsCommand = new Command('params')
+  .description('Handle params')
+  .argument('<value>')
+  .action((value) => {
+    console.log('Received param:', value);
+  });
